tests/submission/MOSQUITO/device-orientation/interval-constant.html

<!DOCTYPE html>
<html>

  <head>
	<title>The interval property must be expressed in milliseconds. It must be a constant, to simplify filtering of the data by the Web application:</title>
	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="http://w3c-test.org/resources/testharness.js"></script>
	<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
	<script src="http://w3c-test.org/resources/WebIDLParser.js"></script>
	<script src="http://w3c-test.org/resources/idlharness.js"></script>
  </head>

  <body>

  
  	<h1>The interval property must be expressed in milliseconds. It must be a constant, to simplify filtering of the data by the Web application:</h1>
	<div id="log">
	</div>

	<script type="text/javascript">
		tint = async_test("The interval property must be different to zero and must be a constant");
		inter1 = 0;
		inter2 = 0;
		window.addEventListener('devicemotion', function (e){
			if(inter1 == 0){
				inter2 = e.interval;
			}
			inter1 = e.interval;
		}, false);
		
		setTimeout("testing()", 400);
		
		function testing(){
			tint.step(function() {assert_true((inter1 != 0) && (inter1 == inter2))});
			tint.done();
		}
	</script>
	
  </body>
  
</html>
